+++ ChangeLog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `e3nn.norm` for `IrrepsArray`
 - `e3nn.tensor_product`
 - `e3nn.normal`
+- Better support of `+ - * /` operators for `IrrepsArray`
 
 ### Removed
 - `IrrepsArray.cat` is now deprecated and replaced by `e3nn.concatenate`